requirements.txt
python-dotenv
httpx
fastapi
uvicorn
pydantic